[frontend] Breaking out classes into multiple files
